Implement word break escape support
Similar to GNU/Readline it is now possible to specify an
array of characters to act as escape chars.

When a word break char is prefixed with an escape char
it won't be recognized as a word break char anymore.

This for example allows you to have spaces in the completion
and hints lists without breaking its functionality.
This requires the lists to be escaped already and it should
be only used one character for escaping.