Add history_move_{next/prevoius}
Right now there is UP/DOWN arrows, that is binded to
history_prevous/history_next accordingly.

But in case of multiline history entry it will navigate through the
lines in the current history itemfirst, and this may not be the requird
behavior.

So now replxx will have history_move_*, that will always navigate
through history, regardless of new lines in query.

New history_move_{next/prev} binded to M-UP/M-DOWN (like in readline).